memperbaiki controller, migration, models, public, routes dan views

+++ migrations/20221222075323-create-company.js
@@ -2,34 +2,34 @@
 /** @type {import('sequelize-cli').Migration} */
 module.exports = {
     async up(queryInterface, Sequelize) {
-        await queryInterface.createTable('Companies', {
-            id: {
-                allowNull: false,
-                autoIncrement: true,
-                primaryKey: true,
-                type: Sequelize.INTEGER
-            },
-            name: {
-                allowNull: false,
-                type: Sequelize.STRING
-            },
-            no_telp: {
-                allowNull: false,
-                type: Sequelize.INTEGER
-            },
-            address: {
-                allowNull: false,
-                type: Sequelize.STRING
-            },
-            createdAt: {
-                allowNull: false,
-                type: Sequelize.DATE
-            },
-            updatedAt: {
-                allowNull: false,
-                type: Sequelize.DATE
-            }
-        });
+        await queryInterface.createTable("Companies", {
+					id: {
+						allowNull: false,
+						autoIncrement: true,
+						primaryKey: true,
+						type: Sequelize.INTEGER,
+					},
+					name: {
+						allowNull: false,
+						type: Sequelize.STRING,
+					},
+					no_telp: {
+						allowNull: false,
+						type: Sequelize.BIGINT,
+					},
+					address: {
+						allowNull: false,
+						type: Sequelize.STRING,
+					},
+					createdAt: {
+						allowNull: false,
+						type: Sequelize.DATE,
+					},
+					updatedAt: {
+						allowNull: false,
+						type: Sequelize.DATE,
+					},
+				});
     },
     async down(queryInterface, Sequelize) {
         await queryInterface.dropTable('Companies');
